move errors to error.rs
